+++ exporter/exporterhelper/internal/batch_sender_test.go
@@ -16,46 +16,49 @@ import (
 
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/exporter/exporterbatcher"
+	"go.opentelemetry.io/collector/exporter/exporterhelper/internal/requesttest"
 	"go.opentelemetry.io/collector/exporter/exporterqueue"
-	"go.opentelemetry.io/collector/exporter/internal"
-	"go.opentelemetry.io/collector/exporter/internal/requesttest"
 )
 
 func TestBatchSender_Merge(t *testing.T) {
 	if runtime.GOOS == "windows" {
 		t.Skip("skipping flaky test on Windows, see https://github.com/open-telemetry/opentelemetry-collector/issues/10758")
 	}
-	cfg := exporterbatcher.NewDefaultConfig()
-	cfg.MinSizeItems = 10
-	cfg.FlushTimeout = 100 * time.Millisecond
 
 	tests := []struct {
-		name          string
-		batcherOption Option
+		name     string
+		batchCfg exporterbatcher.Config
 	}{
 		{
-			name:          "split_disabled",
-			batcherOption: WithBatcher(cfg),
+			name: "split_disabled",
+			batchCfg: func() exporterbatcher.Config {
+				cfg := exporterbatcher.NewDefaultConfig()
+				cfg.MinSizeItems = 10
+				cfg.FlushTimeout = 100 * time.Millisecond
+				return cfg
+			}(),
 		},
 		{
 			name: "split_high_limit",
-			batcherOption: func() Option {
-				c := cfg
-				c.MaxSizeItems = 1000
-				return WithBatcher(c)
+			batchCfg: func() exporterbatcher.Config {
+				cfg := exporterbatcher.NewDefaultConfig()
+				cfg.MinSizeItems = 10
+				cfg.FlushTimeout = 100 * time.Millisecond
+				cfg.MaxSizeItems = 1000
+				return cfg
 			}(),
 		},
 	}
 
 	runTest := func(testName string, enableQueueBatcher bool, tt struct {
-		name          string
-		batcherOption Option
+		name     string
+		batchCfg exporterbatcher.Config
 	},
 	) {
 		t.Run(testName, func(t *testing.T) {
 			resetFeatureGate := setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
-			be := queueBatchExporter(t, tt.batcherOption)
-
+			be, err := newQueueBatchExporter(exporterqueue.NewDefaultConfig(), tt.batchCfg)
+			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 			t.Cleanup(func() {
 				require.NoError(t, be.Shutdown(context.Background()))
@@ -98,32 +101,36 @@ func TestBatchSender_Merge(t *testing.T) {
 }
 
 func TestBatchSender_BatchExportError(t *testing.T) {
-	cfg := exporterbatcher.NewDefaultConfig()
-	cfg.MinSizeItems = 10
 	tests := []struct {
 		name             string
-		batcherOption    Option
+		batchCfg         exporterbatcher.Config
 		expectedRequests int64
 		expectedItems    int64
 	}{
 		{
-			name:          "merge_only",
-			batcherOption: WithBatcher(cfg),
+			name: "merge_only",
+			batchCfg: func() exporterbatcher.Config {
+				cfg := exporterbatcher.NewDefaultConfig()
+				cfg.MinSizeItems = 10
+				return cfg
+			}(),
 		},
 		{
 			name: "merge_without_split_triggered",
-			batcherOption: func() Option {
-				c := cfg
-				c.MaxSizeItems = 200
-				return WithBatcher(c)
+			batchCfg: func() exporterbatcher.Config {
+				cfg := exporterbatcher.NewDefaultConfig()
+				cfg.MinSizeItems = 10
+				cfg.MaxSizeItems = 200
+				return cfg
 			}(),
 		},
 		{
 			name: "merge_with_split_triggered",
-			batcherOption: func() Option {
-				c := cfg
-				c.MaxSizeItems = 20
-				return WithBatcher(c)
+			batchCfg: func() exporterbatcher.Config {
+				cfg := exporterbatcher.NewDefaultConfig()
+				cfg.MinSizeItems = 10
+				cfg.MaxSizeItems = 20
+				return cfg
 			}(),
 			expectedRequests: 1,
 			expectedItems:    20,
@@ -131,16 +138,17 @@ func TestBatchSender_BatchExportError(t *testing.T) {
 	}
 	runTest := func(testName string, enableQueueBatcher bool, tt struct {
 		name             string
-		batcherOption    Option
+		batchCfg         exporterbatcher.Config
 		expectedRequests int64
 		expectedItems    int64
 	},
 	) {
 		t.Run(testName, func(t *testing.T) {
 			resetFeatureGate := setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
-			be := queueBatchExporter(t, tt.batcherOption)
-
+			be, err := newQueueBatchExporter(exporterqueue.NewDefaultConfig(), tt.batchCfg)
+			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
+
 			t.Cleanup(func() {
 				require.NoError(t, be.Shutdown(context.Background()))
 				resetFeatureGate()
@@ -163,7 +171,7 @@ func TestBatchSender_BatchExportError(t *testing.T) {
 			assert.Eventually(t, func() bool {
 				return sink.RequestsCount() == tt.expectedRequests &&
 					sink.ItemsCount() == tt.expectedItems &&
-					be.QueueSender.(*QueueSender).queue.Size() == 0
+					be.queue.Size() == 0
 			}, 100*time.Millisecond, 10*time.Millisecond)
 		})
 	}
@@ -178,13 +186,14 @@ func TestBatchSender_MergeOrSplit(t *testing.T) {
 		t.Run(testName, func(t *testing.T) {
 			resetFeatureGate := setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)
 
-			cfg := exporterbatcher.NewDefaultConfig()
-			cfg.MinSizeItems = 5
-			cfg.MaxSizeItems = 10
-			cfg.FlushTimeout = 100 * time.Millisecond
-			be := queueBatchExporter(t, WithBatcher(cfg))
-
+			batchCfg := exporterbatcher.NewDefaultConfig()
+			batchCfg.MinSizeItems = 5
+			batchCfg.MaxSizeItems = 10
+			batchCfg.FlushTimeout = 100 * time.Millisecond
+			be, err := newQueueBatchExporter(exporterqueue.NewDefaultConfig(), batchCfg)
+			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
+
 			t.Cleanup(func() {
 				require.NoError(t, be.Shutdown(context.Background()))
 				resetFeatureGate()
@@ -229,8 +238,8 @@ func TestBatchSender_Shutdown(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			batchCfg := exporterbatcher.NewDefaultConfig()
 			batchCfg.MinSizeItems = 10
-			be := queueBatchExporter(t, WithBatcher(batchCfg))
-
+			be, err := newQueueBatchExporter(exporterqueue.NewDefaultConfig(), batchCfg)
+			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
 			sink := requesttest.NewSink()
@@ -247,34 +256,6 @@ func TestBatchSender_Shutdown(t *testing.T) {
 		})
 	}
 
-	runTest("enable_queue_batcher", true)
-	runTest("disable_queue_batcher", false)
-}
-
-func TestBatchSender_Disabled(t *testing.T) {
-	runTest := func(testName string, enableQueueBatcher bool) {
-		t.Run(testName, func(t *testing.T) {
-			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
-			cfg := exporterbatcher.NewDefaultConfig()
-			cfg.Enabled = false
-			cfg.MaxSizeItems = 5
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(cfg))
-			require.NotNil(t, be)
-			require.NoError(t, err)
-
-			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
-			t.Cleanup(func() {
-				require.NoError(t, be.Shutdown(context.Background()))
-			})
-
-			sink := requesttest.NewSink()
-			// should be sent right away without splitting because batching is disabled.
-			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 8, Sink: sink}))
-			assert.Equal(t, int64(1), sink.RequestsCount())
-			assert.Equal(t, int64(8), sink.ItemsCount())
-		})
-	}
 	runTest("enable_queue_batcher", true)
 	runTest("disable_queue_batcher", false)
 }
@@ -313,9 +294,7 @@ func TestBatchSender_PostShutdown(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(exporterbatcher.NewDefaultConfig()))
-			require.NotNil(t, be)
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, exporterbatcher.NewDefaultConfig())
 			require.NoError(t, err)
 			assert.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 			assert.NoError(t, be.Shutdown(context.Background()))
@@ -339,13 +318,13 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 	}
 	tests := []struct {
 		name             string
-		batcherCfg       exporterbatcher.Config
+		bCfg             exporterbatcher.Config
 		expectedRequests int64
 		expectedItems    int64
 	}{
 		{
 			name: "merge_only",
-			batcherCfg: func() exporterbatcher.Config {
+			bCfg: func() exporterbatcher.Config {
 				cfg := exporterbatcher.NewDefaultConfig()
 				cfg.FlushTimeout = 20 * time.Millisecond
 				return cfg
@@ -355,7 +334,7 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 		},
 		{
 			name: "merge_without_split_triggered",
-			batcherCfg: func() exporterbatcher.Config {
+			bCfg: func() exporterbatcher.Config {
 				cfg := exporterbatcher.NewDefaultConfig()
 				cfg.FlushTimeout = 20 * time.Millisecond
 				cfg.MaxSizeItems = 200
@@ -366,7 +345,7 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 		},
 		{
 			name: "merge_with_split_triggered",
-			batcherCfg: func() exporterbatcher.Config {
+			bCfg: func() exporterbatcher.Config {
 				cfg := exporterbatcher.NewDefaultConfig()
 				cfg.FlushTimeout = 50 * time.Millisecond
 				cfg.MaxSizeItems = 10
@@ -388,10 +367,7 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			qCfg := exporterqueue.NewDefaultConfig()
 			qCfg.NumConsumers = 2
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(tt.batcherCfg),
-				WithRequestQueue(qCfg, exporterqueue.NewMemoryQueueFactory[internal.Request]()))
-			require.NotNil(t, be)
+			be, err := newQueueBatchExporter(qCfg, tt.bCfg)
 			require.NoError(t, err)
 			assert.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 			t.Cleanup(func() {
@@ -400,8 +376,8 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 
 			sink := requesttest.NewSink()
 			// the 1st and 2nd request should be flushed in the same batched request by max concurrency limit.
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
 
 			assert.Eventually(t, func() bool {
 				return sink.RequestsCount() == 1 && sink.ItemsCount() == 4
@@ -414,25 +390,25 @@ func TestBatchSender_ConcurrencyLimitReached(t *testing.T) {
 			}, 100*time.Millisecond, 10*time.Millisecond)
 
 			// the 4th and 5th request should be flushed in the same batched request by max concurrency limit.
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 2, Sink: sink}))
 			assert.Eventually(t, func() bool {
 				return sink.RequestsCount() == 3 && sink.ItemsCount() == 10
 			}, 100*time.Millisecond, 10*time.Millisecond)
 
 			// do it a few more times to ensure it produces the correct batch size regardless of goroutine scheduling.
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 5, Sink: sink}))
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 6, Sink: sink}))
-			if tt.batcherCfg.MaxSizeItems == 10 {
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 5, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 6, Sink: sink}))
+			if tt.bCfg.MaxSizeItems == 10 {
 				// in case of MaxSizeItems=10, wait for the leftover request to send
 				assert.Eventually(t, func() bool {
 					return sink.RequestsCount() == 5 && sink.ItemsCount() == 21
 				}, 50*time.Millisecond, 10*time.Millisecond)
 			}
 
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 6, Sink: sink}))
-			assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 20, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 6, Sink: sink}))
+			require.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 20, Sink: sink}))
 			assert.Eventually(t, func() bool {
 				return sink.RequestsCount() == tt.expectedRequests && sink.ItemsCount() == tt.expectedItems
 			}, 100*time.Millisecond, 10*time.Millisecond)
@@ -446,13 +422,11 @@ func TestBatchSender_BatchBlocking(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			bCfg := exporterbatcher.NewDefaultConfig()
 			bCfg.MinSizeItems = 3
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender, WithBatcher(bCfg))
-			require.NotNil(t, be)
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
 			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
 			sink := requesttest.NewSink()
-
 			// send 6 blocking requests
 			wg := sync.WaitGroup{}
 			for i := 0; i < 6; i++ {
@@ -482,13 +456,11 @@ func TestBatchSender_BatchCancelled(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			bCfg := exporterbatcher.NewDefaultConfig()
 			bCfg.MinSizeItems = 2
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender, WithBatcher(bCfg))
-			require.NotNil(t, be)
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
 			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
 			sink := requesttest.NewSink()
-
 			// send 2 blocking requests
 			wg := sync.WaitGroup{}
 			ctx, cancel := context.WithCancel(context.Background())
@@ -523,9 +495,8 @@ func TestBatchSender_DrainActiveRequests(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			bCfg := exporterbatcher.NewDefaultConfig()
 			bCfg.MinSizeItems = 2
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(bCfg))
-			require.NotNil(t, be)
+
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
 			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
@@ -561,10 +532,8 @@ func TestBatchSender_UnstartedShutdown(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
 		t.Run(testName, func(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(exporterbatcher.NewDefaultConfig()))
+			be, err := newQueueBatchExporter(exporterqueue.NewDefaultConfig(), exporterbatcher.NewDefaultConfig())
 			require.NoError(t, err)
-
 			err = be.Shutdown(context.Background())
 			require.NoError(t, err)
 		})
@@ -626,22 +595,20 @@ func TestBatchSenderWithTimeout(t *testing.T) {
 			defer setFeatureGateForTest(t, usePullingBasedExporterQueueBatcher, enableQueueBatcher)()
 			bCfg := exporterbatcher.NewDefaultConfig()
 			bCfg.MinSizeItems = 10
-			tCfg := NewDefaultTimeoutConfig()
-			tCfg.Timeout = 50 * time.Millisecond
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(bCfg),
-				WithTimeout(tCfg))
+
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
 			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
+			ctx, cancel := context.WithTimeout(context.Background(), 50*time.Millisecond)
+			defer cancel()
 			sink := requesttest.NewSink()
-
 			// Send 3 concurrent requests that should be merged in one batch
 			wg := sync.WaitGroup{}
 			for i := 0; i < 3; i++ {
 				wg.Add(1)
 				go func() {
-					assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
+					assert.NoError(t, be.Send(ctx, &requesttest.FakeRequest{Items: 4, Sink: sink}))
 					wg.Done()
 				}()
 			}
@@ -653,7 +620,7 @@ func TestBatchSenderWithTimeout(t *testing.T) {
 			for i := 0; i < 3; i++ {
 				wg.Add(1)
 				go func() {
-					assert.Error(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink, Delay: 30 * time.Millisecond}))
+					assert.Error(t, be.Send(ctx, &requesttest.FakeRequest{Items: 4, Sink: sink, Delay: 30 * time.Millisecond}))
 					wg.Done()
 				}()
 			}
@@ -670,50 +637,32 @@ func TestBatchSenderWithTimeout(t *testing.T) {
 	runTest("disable_queue_batcher", false)
 }
 
-// func TestBatchSenderTimerResetNoConflict(t *testing.T) {
-// 	delayBatchMergeFunc := func(_ context.Context, r1 internal.Request, r2 internal.Request) (internal.Request, error) {
-// 		time.Sleep(30 * time.Millisecond)
-// 		if r1 == nil {
-// 			return r2, nil
-// 		}
-// 		fr1 := r1.(*requesttest.FakeRequest)
-// 		fr2 := r2.(*requesttest.FakeRequest)
-// 		if fr2.mergeErr != nil {
-// 			return nil, fr2.mergeErr
-// 		}
-// 		return &requesttest.FakeRequest{
-// 			Items:     fr1.items + fr2.items,
-// 			Sink:      fr1.sink,
-// 			ExportErr: fr2.exportErr,
-// 			Delay:     fr1.delay + fr2.delay,
-// 		}, nil
-// 	}
-// 	bCfg := exporterbatcher.NewDefaultConfig()
-// 	bCfg.MinSizeItems = 8
-// 	bCfg.FlushTimeout = 50 * time.Millisecond
-// 	be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-// 		WithBatcher(bCfg))
-// 	require.NoError(t, err)
-// 	require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
-// 	sink := requesttest.NewSink()
+func TestBatchSenderTimerResetNoConflict(t *testing.T) {
+	bCfg := exporterbatcher.NewDefaultConfig()
+	bCfg.MinSizeItems = 8
+	bCfg.FlushTimeout = 100 * time.Millisecond
+	be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
+	require.NoError(t, err)
+	require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 
-// 	// Send 2 concurrent requests that should be merged in one batch in the same interval as the flush timer
-// 	go func() {
-// 		assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
-// 	}()
-// 	time.Sleep(30 * time.Millisecond)
-// 	go func() {
-// 		assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
-// 	}()
+	sink := requesttest.NewSink()
+	// Send 2 concurrent requests that should be merged in one batch in the same interval as the flush timer
+	go func() {
+		assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
+	}()
+	time.Sleep(30 * time.Millisecond)
+	go func() {
+		assert.NoError(t, be.Send(context.Background(), &requesttest.FakeRequest{Items: 4, Sink: sink}))
+	}()
 
-// 	// The batch should be sent either with the flush interval or by reaching the minimum items size with no conflict
-// 	assert.EventuallyWithT(t, func(c *assert.CollectT) {
-// 		assert.LessOrEqual(c, int64(1), sink.RequestsCount())
-// 		assert.EqualValues(c, 8, sink.ItemsCount())
-// 	}, 200*time.Millisecond, 10*time.Millisecond)
+	// The batch should be sent either with the flush interval or by reaching the minimum items size with no conflict
+	assert.EventuallyWithT(t, func(c *assert.CollectT) {
+		assert.LessOrEqual(c, int64(1), sink.RequestsCount())
+		assert.EqualValues(c, 8, sink.ItemsCount())
+	}, 500*time.Millisecond, 10*time.Millisecond)
 
-// 	require.NoError(t, be.Shutdown(context.Background()))
-// }
+	require.NoError(t, be.Shutdown(context.Background()))
+}
 
 func TestBatchSenderTimerFlush(t *testing.T) {
 	runTest := func(testName string, enableQueueBatcher bool) {
@@ -725,8 +674,7 @@ func TestBatchSenderTimerFlush(t *testing.T) {
 			bCfg := exporterbatcher.NewDefaultConfig()
 			bCfg.MinSizeItems = 8
 			bCfg.FlushTimeout = 100 * time.Millisecond
-			be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender,
-				WithBatcher(bCfg))
+			be, err := newQueueBatchExporter(exporterqueue.Config{}, bCfg)
 			require.NoError(t, err)
 			require.NoError(t, be.Start(context.Background(), componenttest.NewNopHost()))
 			sink := requesttest.NewSink()
@@ -763,12 +711,4 @@ func TestBatchSenderTimerFlush(t *testing.T) {
 	}
 	runTest("enable_queue_batcher", true)
 	runTest("disable_queue_batcher", false)
-}
-
-func queueBatchExporter(t *testing.T, opts ...Option) *BaseExporter {
-	opts = append(opts, WithRequestQueue(exporterqueue.NewDefaultConfig(), exporterqueue.NewMemoryQueueFactory[internal.Request]()))
-	be, err := NewBaseExporter(defaultSettings, defaultSignal, newNoopObsrepSender, opts...)
-	require.NotNil(t, be)
-	require.NoError(t, err)
-	return be
 }