+++ processor/xprocessor/processor.go
@@ -39,6 +39,7 @@ type Profiles interface {
 type CreateProfilesFunc func(context.Context, processor.Settings, component.Config, xconsumer.Profiles) (Profiles, error)
 
 // CreateProfiles implements Factory.CreateProfiles.
+// Deprecated: [v0.120.0] No longer used, will be removed.
 func (f CreateProfilesFunc) CreateProfiles(ctx context.Context, set processor.Settings, cfg component.Config, next xconsumer.Profiles) (Profiles, error) {
 	if f == nil {
 		return nil, pipeline.ErrSignalNotSupported
@@ -61,12 +62,19 @@ func (f factoryOptionFunc) applyOption(o *factoryOpts) {
 
 type factory struct {
 	processor.Factory
-	CreateProfilesFunc
+	createProfilesFunc     CreateProfilesFunc
 	profilesStabilityLevel component.StabilityLevel
 }
 
 func (f factory) ProfilesStability() component.StabilityLevel {
 	return f.profilesStabilityLevel
+}
+
+func (f factory) CreateProfiles(ctx context.Context, set processor.Settings, cfg component.Config, next xconsumer.Profiles) (Profiles, error) {
+	if f.createProfilesFunc == nil {
+		return nil, pipeline.ErrSignalNotSupported
+	}
+	return f.createProfilesFunc(ctx, set, cfg, next)
 }
 
 type factoryOpts struct {
@@ -99,7 +107,7 @@ func WithLogs(createLogs processor.CreateLogsFunc, sl component.StabilityLevel) 
 func WithProfiles(createProfiles CreateProfilesFunc, sl component.StabilityLevel) FactoryOption {
 	return factoryOptionFunc(func(o *factoryOpts) {
 		o.profilesStabilityLevel = sl
-		o.CreateProfilesFunc = createProfiles
+		o.createProfilesFunc = createProfiles
 	})
 }